test-requirements.txt
-r requirements.txt

nose==1.3.3
factory-boy==2.4.1
